デプロイ先heroku変更 docker-compose変更 環境変数 .envローカル追加

+++ BookManagementSystem_app/models.py
@@ -4,11 +4,10 @@ from django.db import models
 class Book(models.Model):
     title = models.CharField(max_length=255)
     genre = models.ForeignKey('Genre', on_delete=models.CASCADE)
-    # user = models.ForeignKey('auth.User', on_delete=models.CASCADE, null=True)
     user = models.CharField(max_length=255, null = True)
     isborrow = models.BooleanField(default=False)
-    num1 = models.IntegerField(default=0)
-    num2 = models.IntegerField(default=0)
+    num1 = models.BigIntegerField(default=0)
+    num2 = models.BigIntegerField(default=0)
 
     def __str__(self):
         return self.title
